Use different number of operations per invocation in memory and rocks benchmarks
